# spec/openapi/uai-v1.yaml
openapi: 3.1.0
info:
  title: Universal Agent Identity API
  version: "0.1.0"
  summary: Identity, attestation and governance for autonomous AI agents.
  description: |
    The UAI v1 REST surface.

    REST and JSON because a protocol meant to be implemented by everyone must be
    implementable with `curl`. gRPC is used internally where call volume justifies it.

    **Authentication.** Every state-changing call requires proof of possession: mutual TLS
    with a SPIFFE X.509-SVID *and* an RFC 9421 HTTP message signature. mTLS proves the
    runtime instance; the message signature proves the persistent identity. Either alone is
    insufficient, because a stolen SVID has no persistent key and a stolen persistent key has
    no attested runtime.

    **What this API does not claim.** A verified identity is not a safe agent, and revocation
    does not stop code from running outside the ecosystem. See
    `docs/protocol/01-overview.md`.
  license:
    name: Apache-2.0 (implementation) / CC-BY-4.0 (specification)
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: UAI security
    email: security@uai.world
servers:
  - url: https://api.uai.world/v1
    description: Production API
  - url: https://verify.uai.world/v1
    description: Public verification endpoint (unauthenticated, cached)
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - mutualTLS: []
    messageSignature: []
paths:
...
  /agents/{id}:
    get:
      tags: [identity]
      operationId: getAgent
      summary: Universal Agent Identity Card
      parameters: [{ $ref: '#/components/parameters/AgentId' }]
      responses:
        '200':
          description: Identity card
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegisteredAgent' }
        '404': { $ref: '#/components/responses/Problem' }
...
  /actions/{eventId}:
    get:
      tags: [action]
      operationId: getAction
      summary: An attestation with its transparency receipt
      parameters:
        - name: eventId
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$' }
      responses:
        '200':
          description: Attestation and proof
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttestationReceipt' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    mutualTLS:
      type: mutualTLS
      description: SPIFFE X.509-SVID. Proves the runtime instance.
    messageSignature:
      type: apiKey
      in: header
      name: Signature
      description: |
        RFC 9421 HTTP message signature over @method, @target-uri, content-digest,
        uai-agent-id and uai-nonce, tagged with the UAI domain. Proves the persistent
        identity. `agent_id` in a JSON body proves nothing.
    webAuthn:
      type: apiKey
      in: header
      name: X-UAI-Assertion
      description: |
        WebAuthn assertion from a hardware authenticator with user verification required.
        For votes, the challenge is the digest of the vote statement itself.
  parameters:
    AgentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^uai:agent:[0-9A-HJKMNP-TV-Z]{26}$'
...
  responses:
    Problem:
      description: RFC 9457 problem document
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: |
        RFC 9457. Carries `decision_id` and `policy_version` whenever a policy was involved, so
        a refusal is auditable and reproducible rather than an opaque denial.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string, examples: [UAI_IDENTITY_REVOKED, UAI_POP_REQUIRED, UAI_CHAIN_CONFLICT] }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        decision_id: { type: string }
        policy_version: { type: string }
        governance_proof: { type: string }
        remediation: { type: string }
        trace_id: { type: string }
...
    RegisteredAgent:
      type: object
      required: [uai_id, did, status]
      properties:
        uai_id: { type: string }
        did: { type: string }
        logical_name: { type: string }
        status:
          enum: [UNREGISTERED, REGISTERED, VERIFIED, ACTIVE, UNBOUND, QUARANTINED, REVOCATION_AUTHORIZED, REVOKED]
        assurance_level: { enum: [UAI-AL0, UAI-AL1, UAI-AL2, UAI-AL3] }
        identity_commitment: { type: string }
        credentials:
          type: array
          items: { type: string }
          description: Types of the credentials issued with the identity (§8.2).
        genesis_event_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
          description: |
            Anchor of the agent's event chain. The first attestation references it as
            previous_event_hash, so a verifier walking backwards from any action reaches
            registration through an unbroken hash path.
        policy_version: { type: string }
        transparency: { type: object }
...
    AttestationReceipt:
      type: object
      required: [event_id]
      properties:
        event_id: { type: string }
        attestation: { $ref: '#/components/schemas/ActionAttestation' }
        receipt: { $ref: '#/components/schemas/TransparencyReceipt' }
...
    ActionAttestation:
      $ref: '../schemas/action-attestation.schema.json'
...
    TransparencyReceipt:
      $ref: '../schemas/transparency-receipt.schema.json'
...